+++ src/ResourceManager/Cdn/Commands.Cdn/Endpoint/TestAzureRmCdnCustomDomain.cs
@@ -23,7 +23,7 @@ using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 namespace Microsoft.Azure.Commands.Cdn.Endpoint
 {
     [Cmdlet("Test", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CdnCustomDomain", DefaultParameterSetName = FieldsParameterSet), OutputType(typeof(PSValidateCustomDomainOutput))]
-    [Alias("Validate-AzureRmCdnCustomDomain")]
+    [Alias("Validate-" + ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CdnCustomDomain")]
     public class TestAzureRmCdnCustomDomain : AzureCdnCmdletBase
     {
         [Parameter(Mandatory = true, ParameterSetName = FieldsParameterSet, HelpMessage = "Azure CDN endpoint name.")]